backend/requirements.txt
fastapi
uvicorn[standard]
pyyaml
python-multipart